refactor : 매장 별 쿠폰 통계 cursor 공통 dto 로 이동
