+++ src/main/java/com/sparta/couponpop/domain/couponevent/controller/CouponEventInternalController.java
@@ -9,5 +9,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class CouponEventInternalController {
 
+//    @GetMapping("/")
+//    StoreOwnershipResponse checkOwnership(Long storeId, Long memberId);
+
+//    @GetMapping("/")
+//    List<StoreResponse> findStoresByOwner(Long memberId, StoreCouponEventsStatisticsCursor cursor, int pageSize);
+
+//    @GetMapping("/")
+//    StoreResponse findByIdOrElseThrow(Long storeId);
+
+//    @GetMapping("/")
+//    List<StoreResponse> findAllByIds(List<Long> storeIds);
+    // 문서화 필요
 
 }
