+++ src/main/java/com/sparta/couponpop/domain/couponevent/repository/CouponEventQueryRepository.java
@@ -1,7 +1,6 @@
 package com.sparta.couponpop.domain.couponevent.repository;
 
 import com.sparta.couponpop.domain.couponevent.dto.cursor.StoreCouponEventsCursor;
-import com.sparta.couponpop.domain.couponevent.dto.cursor.StoreCouponEventsStatisticsCursor;
 import com.sparta.couponpop.domain.couponevent.enums.CouponEventStatus;
 import com.sparta.couponpop.domain.couponevent.repository.dto.CouponEventWithUsedCountProjection;
 import com.sparta.couponpop.domain.couponevent.repository.dto.StoreCouponEventStatisticsProjection;
