+++ src/main/java/com/sparta/couponpop/domain/couponevent/repository/CouponEventQueryRepositoryImpl.java
@@ -7,14 +7,12 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.sparta.couponpop.domain.coupon.entity.QCoupon;
 import com.sparta.couponpop.domain.coupon.enums.CouponStatus;
 import com.sparta.couponpop.domain.couponevent.dto.cursor.StoreCouponEventsCursor;
-import com.sparta.couponpop.domain.couponevent.dto.cursor.StoreCouponEventsStatisticsCursor;
 import com.sparta.couponpop.domain.couponevent.entity.QCouponEvent;
 import com.sparta.couponpop.domain.couponevent.enums.CouponEventStatus;
 import com.sparta.couponpop.domain.couponevent.repository.dto.*;
 import com.sparta.couponpop.domain.store.entity.QStore;
 import com.sparta.couponpop.domain.couponevent.repository.dto.CouponEventWithUsedCountProjection;
 import com.sparta.couponpop.domain.couponevent.repository.dto.QCouponEventWithUsedCountProjection;
-import com.sparta.couponpop.domain.store.entity.Store;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.util.ObjectUtils;
